requirements.txt
# Core dependencies for AI Playwright Testing Engine
# These are the essential packages needed for basic functionality

# Web automation and testing
playwright==1.40.0

# Async and web framework
aiohttp==3.9.1
fastapi==0.104.1
uvicorn==0.24.0

# Data validation and serialization
pydantic==2.5.0

# Database (SQLite by default, PostgreSQL optional)
sqlalchemy==2.0.23

# Basic data science and ML
numpy>=1.26.0,<2.0.0  # Python 3.12 compatible
pandas==2.0.3
scikit-learn==1.3.2

# Web scraping and parsing
beautifulsoup4==4.12.2

# System monitoring
psutil==5.9.6

# Configuration and utilities
pyyaml==6.0.1
python-dotenv==1.0.0

# Logging
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Basic plotting (lightweight)
matplotlib==3.7.2

# Note: Some packages removed to avoid compilation issues:
# - tensorflow (requires C++ compiler)
# - opencv-python (requires additional system libraries)
# - plotly (can be added later if needed)
# - psycopg2-binary (only needed for PostgreSQL)
# - celery/redis (only needed for distributed processing)
